refactor: extract shared multi-origin palette lookup helpers
Add resolvePresetColorBySlug / resolvePresetColorByHex in src/utils/
that walk origin groups in reverse so custom palette wins over
theme/default when slugs or hex values collide, mirroring how
--wp--preset--color--{slug} CSS variables resolve at render time.

Apply the helpers to:
- src/utils/encode-color-value.js: fixes the same first-match-wins
  collision bug in the shared encode/decode utilities used by
  ~20 blocks (text-reveal, expanding-bg, svg-patterns, card,
  section, etc.) so the swatch highlighted in
  ColorGradientSettingsDropdown matches the CSS variable cascade.
- src/overlay-header/index.js: replace the inline lookup helpers
  introduced in the previous commit.
